agent-factory/requirements.txt
# Google ADK
google-adk>=1.2.0

# Authentication for Cloud Run
google-auth>=2.34.0
google-auth-httplib2>=0.2.0

# Google Cloud Services
google-cloud-aiplatform>=1.60.0

# HTTP Client for MCP
httpx>=0.27.0

# Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Utilities
python-dotenv>=1.0.0

toolbox-core